==> nbs/01_Cleaning_Data.ipynb <==
# Cleaning Data

#| hide
import pandas as pd
import numpy as np

We will be using the `Concatenated_Orig_data.csv`, that was created using the `00_Collecting_Data.ipynb`

df = pd.read_csv(f"../data/Concatenated_Orig_data.csv",encoding='latin-1', header=0)

## Preprocessing

First, lets examine our data:

df.head()

Let's clean the data set by dropping duplicate entries by the identification number and converting blank cells to `NaN`

df = df.drop_duplicates(subset=['c5'])
df.replace(r'^\s*$', np.nan, regex=True, inplace=True)

## Selecting Opcodes

#| hide
import numpy as np 
import pandas as pd 
import matplotlib.pyplot as plt 
import seaborn as sns

Now, let's get a better sense of the target column `c78`. Specifcally, we'd like to know how many instances there are of each label, and limit ourselves to only the most frequent labels to reduce numerosity.

Let's limit the data to those entries with the value of `c78` appearing in the `AIDCODE.csv`

df2 = pd.read_csv("../data/raw-data/AIDCODE.csv")
list_code = df2['CODE'].unique()
list_code = list_code.tolist()

#drop rows that contain any value in the list
df = df[df.c78.isin(list_code) == True]

df.dropna(subset=['c78'],inplace=True)

df['c78'].value_counts()

df.shape[0]

## Saving

df.to_csv("../data/Concatenated_Clean_data.csv")
