==> nbs/07_example.ipynb <==
# Example

Here's a single example of our workflow using the NN model

Let's pull a random entry from our datasets

text = 'TOOK OFF WITH ONE GENERATOR INOPERATIVE, THE OTHER THEN FAILED, CAUSING INSUFFICIENT POWER TO LOCK LANDING GEAR.'
label = 'AU'
label_encoded = 2

This text description corresponds with lablel `AU` which is encoded as `2`

First we tokenize our imput

from transformers import AutoTokenizer

model_path = '../model/'
tokenizer = AutoTokenizer.from_pretrained(model_path)

inputs = tokenizer(text, return_tensors='pt')

Now we pass the inputs to the model

from transformers import AutoModelForSequenceClassification
import torch

model = AutoModelForSequenceClassification.from_pretrained(model_path)
with torch.no_grad():
    logits = model(**inputs).logits

Lets check our results

logits.argmax().item()

As we can see our model is incorrect
